backend-webapp/DBS-Vertex-AI/requirement.txt
Flask==2.1.0
google-cloud-aiplatform==1.25.0
better-profanity==0.7.0